Progress bar

+++ getsrc.py
@@ -7,10 +7,12 @@ from webdriver_manager.chrome import ChromeDriverManager
 import os
 import sys
 from selenium.webdriver.common.by import By
+from selenium.webdriver.chrome.service import Service
 from selenium import webdriver
 import pandas as pd
 import json
-
+from colorama import Fore
+from tqdm import tqdm
 ''' 
 STEP ONE:
 Scrape images from the URL
@@ -18,13 +20,13 @@ Sample usage:
 python getsrc.py https://google.com/
 
 '''
-
 page_data = {}
 
 BYTE_SIZE = 1024
 PHONE_WIDTH = 360
 PHONE_HEIGHT = 640
 PIXEL_RATIO = 3.0
+PROGRESS_BAR = "{l_bar}%s{bar}%s{r_bar}" % (Fore.GREEN, Fore.RESET)
 # Set options for web scraper
 options = webdriver.ChromeOptions()
 options.add_argument('--ignore-certificate-errors')
@@ -42,7 +44,7 @@ mobile_emulation = {
 options.add_experimental_option("mobileEmulation", mobile_emulation)
 
 # Chrome web driver
-driver = webdriver.Chrome(ChromeDriverManager().install(), options=options,desired_capabilities=caps)
+driver = webdriver.Chrome(service=Service(ChromeDriverManager().install()), options=options,desired_capabilities=caps)
 
 
 # URL settings, to be replaced with text file of URLs (possible)
@@ -62,27 +64,27 @@ driver.get(url)
 # Get all elements labelled 'img'
 images = driver.find_elements(By.TAG_NAME, 'img')
 
-print("===== WEBPAGE SIZE =====")
+print("===== GETTING WEBPAGE SIZE =====")
 total_bytes = []
-for entry in driver.get_log('performance'):
-        if "Network.dataReceived" in str(entry):
-            r = re.search(r'encodedDataLength\":(.*?),', str(entry))
-            total_bytes.append(int(r.group(1)))
-            kb = round((float(sum(total_bytes) / BYTE_SIZE)), 2)
+for entry in tqdm(driver.get_log('performance'), bar_format=PROGRESS_BAR):
+    if "Network.dataReceived" in str(entry):
+        r = re.search(r'encodedDataLength\":(.*?),', str(entry))
+        total_bytes.append(int(r.group(1)))
+        kb = round((float(sum(total_bytes) / BYTE_SIZE)), 2)
 print(kb, "KBs")
 page_data["kiloBytesIn"] = kb
 
 # Make directory for this website, in case it doesn't exist. If this gives you a syntax error then reformat the website name.
 try:
-    os.system("mkdir " + host)
+    os.system("mkdir -p " + host)
 except:
     pass
 
 num_img = 0
 results = pd.DataFrame(columns=("Image Source", "Image Name", "Original Size (KB)"))
-print("===== IMAGES =====")
+print("===== SCRAPING IMAGES =====")
 with open(host+"/images.txt", "w") as f:
-    for image in images:
+    for image in tqdm(images, bar_format=PROGRESS_BAR):
         i = image.get_attribute('src')
         try:
             # Write to images
@@ -93,8 +95,6 @@ with open(host+"/images.txt", "w") as f:
             image_name = i.split("/")[-1]
             # Get original image size
             img_size = int(meta.get(name="Content-Length"))/BYTE_SIZE
-            print(img_size, "KBs")
-            
             # Add data to results
             results.loc[num_img] = [i, image_name, img_size]
             num_img+=1
@@ -108,10 +108,13 @@ TO DO:
 - Use wget to download all images from the images.txt file into the directory of the host
 - Possible command (for linux):
 '''
-os.system("cd " + host + " && wget -i images.txt")
+print("===== DOWNLOADING IMAGES =====")
+os.system("cd " + host + " && wget -q --show-progress -i images.txt")
+
 image_names = os.listdir(host)
 for image in image_names:
-    print(image)
+    if image == "page_data.json" or image == "results.csv" or image == "images.txt":
+        continue
     os.system("cd " + host + " && magick " +image +" -define webp:lossless=true " + image.split(".")[0] + ".webp && rm " +image)
 
 
@@ -123,3 +126,5 @@ f.close()
 results.to_csv(host+"/results.csv")
 
 driver.close()
+
+print("===== GET SRC COMPLETE =====")
